Fix race condition in tests
Fix a race condition in the parent/child test, where the parent
process could potentially read form the child’s input file descriptor
(the readable end of the pipe) before the child has a chance to.

Destroying the `net.Socket` instance directly after passing the FD
to the child fixes this.